Separated tests into individual test files
