+++ tests/e2e/events_test/events_test.go
@@ -11,7 +11,6 @@ import (
 	"time"
 
 	"github.com/HackIllinois/api/common/configloader"
-	checkin_models "github.com/HackIllinois/api/services/checkin/models"
 	event_models "github.com/HackIllinois/api/services/event/models"
 	profile_models "github.com/HackIllinois/api/services/profile/models"
 	"github.com/HackIllinois/api/tests/common"
@@ -216,1279 +215,8 @@ func ClearEvents() {
 	client.Database(events_db_name).Collection("favorites").DeleteMany(context.Background(), bson.D{})
 }
 
-func TestGetEventNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	event_id := "testeventid12345"
-	received_event := event_models.Event{}
-	response, err := public_client.New().Get(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_event := event_models.Event{
-		ID:          "testeventid12345",
-		Name:        "testevent1",
-		Description: "testdescription1",
-		StartTime:   current_unix_time,
-		EndTime:     current_unix_time + 60000,
-		Sponsor:     "testsponsor1",
-		EventType:   "WORKSHOP",
-		Locations: []event_models.EventLocation{
-			{
-				Description: "testlocationdescription1",
-				Tags:        []string{"SIEBEL3", "ECEB2"},
-				Latitude:    123.456,
-				Longitude:   123.456,
-			},
-		},
-		Points: 50,
-	}
-
-	if !reflect.DeepEqual(received_event, expected_event) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", expected_event, received_event)
-	}
-}
-
-func TestGetEventNotExist(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	event_id := "nonsense_eventid"
-	received_event := event_models.Event{}
-	response, err := public_client.New().Get(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Error("Unable to make request")
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Errorf("Request returned HTTP error %d", response.StatusCode)
-	}
-}
-
-func TestDeleteNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	event_id := "testeventid12345"
-	received_event := event_models.Event{}
-	response, err := staff_client.New().Delete(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_event := event_models.Event{
-		ID:          "testeventid12345",
-		Name:        "testevent1",
-		Description: "testdescription1",
-		StartTime:   current_unix_time,
-		EndTime:     current_unix_time + 60000,
-		Sponsor:     "testsponsor1",
-		EventType:   "WORKSHOP",
-		Locations: []event_models.EventLocation{
-			{
-				Description: "testlocationdescription1",
-				Tags:        []string{"SIEBEL3", "ECEB2"},
-				Latitude:    123.456,
-				Longitude:   123.456,
-			},
-		},
-		Points: 50,
-	}
-
-	if !reflect.DeepEqual(received_event, expected_event) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", expected_event, received_event)
-	}
-
-	response, err = staff_client.New().Get(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(nil)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestDeleteNotExist(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	event_id := "nonsense_eventid"
-	received_event := event_models.Event{}
-	response, err := staff_client.New().Delete(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	cursor, _ := client.Database(events_db_name).Collection("events").Find(context.Background(), bson.D{})
-	res := []event_models.Event{}
-	err = cursor.All(context.TODO(), &res)
-
-	if err != nil {
-		t.Fatalf("Test failed due to unexpected error: %v", err)
-		return
-	}
-
-	expected_events := []event_models.Event{
-		{
-			ID:          "testeventid12345",
-			Name:        "testevent1",
-			Description: "testdescription1",
-			StartTime:   current_unix_time,
-			EndTime:     current_unix_time + 60000,
-			Sponsor:     "testsponsor1",
-			EventType:   "WORKSHOP",
-			Locations: []event_models.EventLocation{
-				{
-					Description: "testlocationdescription1",
-					Tags:        []string{"SIEBEL3", "ECEB2"},
-					Latitude:    123.456,
-					Longitude:   123.456,
-				},
-			},
-			Points: 50,
-		},
-		{
-			ID:          "testeventid67890",
-			Name:        "testevent2",
-			Description: "testdescription2",
-			StartTime:   current_unix_time,
-			EndTime:     current_unix_time + 60000,
-			Sponsor:     "testsponsor2",
-			EventType:   "WORKSHOP",
-			Locations: []event_models.EventLocation{
-				{
-					Description: "testlocationdescription2",
-					Tags:        []string{"SIEBEL3", "ECEB2"},
-					Latitude:    123.456,
-					Longitude:   123.456,
-				},
-			},
-			Points: 100,
-		},
-	}
-
-	if !reflect.DeepEqual(res, expected_events) {
-		t.Fatalf("Database contained wrong event info. Expected %v, got %v", expected_events, res)
-	}
-}
-
-func TestDeleteForbidden(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	event_id := "testeventid12345"
-	received_event := event_models.Event{}
-	response, err := user_client.New().Delete(fmt.Sprintf("/event/%s/", event_id)).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusForbidden {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	cursor, _ := client.Database(events_db_name).Collection("events").Find(context.Background(), bson.D{})
-	res := []event_models.Event{}
-	err = cursor.All(context.TODO(), &res)
-
-	if err != nil {
-		t.Fatalf("Test failed due to unexpected error: %v", err)
-		return
-	}
-
-	expected_events := []event_models.Event{
-		{
-			ID:          "testeventid12345",
-			Name:        "testevent1",
-			Description: "testdescription1",
-			StartTime:   current_unix_time,
-			EndTime:     current_unix_time + 60000,
-			Sponsor:     "testsponsor1",
-			EventType:   "WORKSHOP",
-			Locations: []event_models.EventLocation{
-				{
-					Description: "testlocationdescription1",
-					Tags:        []string{"SIEBEL3", "ECEB2"},
-					Latitude:    123.456,
-					Longitude:   123.456,
-				},
-			},
-			Points: 50,
-		},
-		{
-			ID:          "testeventid67890",
-			Name:        "testevent2",
-			Description: "testdescription2",
-			StartTime:   current_unix_time,
-			EndTime:     current_unix_time + 60000,
-			Sponsor:     "testsponsor2",
-			EventType:   "WORKSHOP",
-			Locations: []event_models.EventLocation{
-				{
-					Description: "testlocationdescription2",
-					Tags:        []string{"SIEBEL3", "ECEB2"},
-					Latitude:    123.456,
-					Longitude:   123.456,
-				},
-			},
-			Points: 100,
-		},
-	}
-
-	if !reflect.DeepEqual(res, expected_events) {
-		t.Fatalf("Database contained wrong event info. Expected %v, got %v", expected_events, res)
-	}
-}
-
-func TestGetAllEventsNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	received_events := event_models.EventList{}
-	response, err := public_client.New().Get("/event/").ReceiveSuccess(&received_events)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_events := event_models.EventList{
-		Events: []event_models.Event{
-			{
-				ID:          "testeventid12345",
-				Name:        "testevent1",
-				Description: "testdescription1",
-				StartTime:   current_unix_time,
-				EndTime:     current_unix_time + 60000,
-				Sponsor:     "testsponsor1",
-				EventType:   "WORKSHOP",
-				Locations: []event_models.EventLocation{
-					{
-						Description: "testlocationdescription1",
-						Tags:        []string{"SIEBEL3", "ECEB2"},
-						Latitude:    123.456,
-						Longitude:   123.456,
-					},
-				},
-				Points: 50,
-			},
-			{
-				ID:          "testeventid67890",
-				Name:        "testevent2",
-				Description: "testdescription2",
-				StartTime:   current_unix_time,
-				EndTime:     current_unix_time + 60000,
-				Sponsor:     "testsponsor2",
-				EventType:   "WORKSHOP",
-				Locations: []event_models.EventLocation{
-					{
-						Description: "testlocationdescription2",
-						Tags:        []string{"SIEBEL3", "ECEB2"},
-						Latitude:    123.456,
-						Longitude:   123.456,
-					},
-				},
-				Points: 100,
-			},
-		},
-	}
-
-	if !reflect.DeepEqual(received_events, expected_events) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", expected_events, received_events)
-	}
-}
-
-func TestGetAllEventsNone(t *testing.T) {
-	received_events := event_models.EventList{}
-	response, err := staff_client.New().Get("/event/").ReceiveSuccess(&received_events)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_events := event_models.EventList{
-		Events: []event_models.Event{},
-	}
-
-	if !reflect.DeepEqual(received_events, expected_events) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", expected_events, received_events)
-	}
-}
-
-func TestGetFilteredEventsNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	received_events := event_models.EventList{}
-	response, err := public_client.New().Get("/event/filter/?name=testevent1").ReceiveSuccess(&received_events)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_events := event_models.EventList{
-		Events: []event_models.Event{
-			{
-				ID:          "testeventid12345",
-				Name:        "testevent1",
-				Description: "testdescription1",
-				StartTime:   current_unix_time,
-				EndTime:     current_unix_time + 60000,
-				Sponsor:     "testsponsor1",
-				EventType:   "WORKSHOP",
-				Locations: []event_models.EventLocation{
-					{
-						Description: "testlocationdescription1",
-						Tags:        []string{"SIEBEL3", "ECEB2"},
-						Latitude:    123.456,
-						Longitude:   123.456,
-					},
-				},
-				Points: 50,
-			},
-		},
-	}
-
-	if !reflect.DeepEqual(received_events, expected_events) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", expected_events, received_events)
-	}
-}
-
-func TestGetFilteredEventsBadArgs(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	response, err := public_client.New().Get("/event/filter/?nonsensefield=trydecipheringthis!").ReceiveSuccess(nil)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestCreateEventNormal(t *testing.T) {
-	defer ClearEvents()
-	event_info := event_models.Event{
-		Name:        "testevent1",
-		Description: "testdescription1",
-		StartTime:   current_unix_time,
-		EndTime:     current_unix_time + 60000,
-		Sponsor:     "testsponsor1",
-		EventType:   "WORKSHOP",
-		Locations: []event_models.EventLocation{
-			{
-				Description: "testlocationdescription1",
-				Tags:        []string{"SIEBEL3", "ECEB2"},
-				Latitude:    123.456,
-				Longitude:   123.456,
-			},
-		},
-		Points: 50,
-	}
-	received_event := event_models.Event{}
-	response, err := staff_client.New().Post("/event/").BodyJSON(event_info).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	event_info.ID = received_event.ID
-
-	if !reflect.DeepEqual(received_event, event_info) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", event_info, received_event)
-	}
-
-	cursor, _ := client.Database(events_db_name).Collection("events").Find(context.Background(), bson.D{})
-	res := []event_models.Event{}
-	err = cursor.All(context.TODO(), &res)
-
-	if err != nil {
-		t.Fatalf("Test failed due to unexpected error: %v", err)
-		return
-	}
-
-	expected_res := []event_models.Event{
-		event_info,
-	}
-
-	if !reflect.DeepEqual(res, expected_res) {
-		t.Fatalf("Database contained wrong event info. Expected %v, got %v", expected_res, res)
-	}
-}
-
-func TestCreateEventForbidden(t *testing.T) {
-	defer ClearEvents()
-	event_info := event_models.Event{
-		Name:        "testevent1",
-		Description: "testdescription1",
-		StartTime:   current_unix_time,
-		EndTime:     current_unix_time + 60000,
-		Sponsor:     "testsponsor1",
-		EventType:   "WORKSHOP",
-		Locations: []event_models.EventLocation{
-			{
-				Description: "testlocationdescription1",
-				Tags:        []string{"SIEBEL3", "ECEB2"},
-				Latitude:    123.456,
-				Longitude:   123.456,
-			},
-		},
-		Points: 50,
-	}
-	received_event := event_models.Event{}
-	response, err := user_client.New().Post("/event/").BodyJSON(event_info).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusForbidden {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	cursor, _ := client.Database(events_db_name).Collection("events").Find(context.Background(), bson.D{})
-	res := []event_models.Event{}
-	err = cursor.All(context.TODO(), &res)
-
-	if err != nil {
-		t.Fatalf("Test failed due to unexpected error: %v", err)
-		return
-	}
-
-	expected_res := []event_models.Event{}
-
-	if !reflect.DeepEqual(res, expected_res) {
-		t.Fatalf("Database contained wrong event info. Expected %v, got %v", expected_res, res)
-	}
-}
-
-func TestUpdateEventNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	res := client.Database(events_db_name).Collection("events").FindOne(context.Background(), bson.M{"id": "testeventid12345"})
-
-	event := event_models.Event{}
-
-	res.Decode(&event)
-
-	event.Description = "It's a new description!"
-	var received_event event_models.Event
-
-	response, err := staff_client.New().Put("/event/").BodyJSON(event).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	if !reflect.DeepEqual(received_event, event) {
-		t.Fatalf("Wrong event info. Expected %v, got %v", event, received_event)
-	}
-}
-
-func TestUpdateEventForbidden(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	res := client.Database(events_db_name).Collection("events").FindOne(context.Background(), bson.M{"id": "testeventid12345"})
-
-	event := event_models.Event{}
-
-	res.Decode(&event)
-
-	event.Description = "It's a new description!"
-	var received_event event_models.Event
-
-	response, err := user_client.New().Put("/event/").BodyJSON(event).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusForbidden {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestUpdateEventNotFound(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	res := client.Database(events_db_name).Collection("events").FindOne(context.Background(), bson.M{"id": "testeventid12345"})
-
-	event := event_models.Event{}
-
-	res.Decode(&event)
-
-	event.ID = "adifferentidthatsnotbeenadded"
-	event.Description = "It's a new description!"
-	var received_event event_models.Event
-
-	response, err := staff_client.New().Put("/event/").BodyJSON(event).ReceiveSuccess(&received_event)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestGetEventCodeNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_code := event_models.EventCode{}
-	id := "testeventid12345"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/code/%s/", id)).ReceiveSuccess(&recieved_code)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_code := event_models.EventCode{
-		ID:         "testeventid12345",
-		Code:       "123456",
-		Expiration: current_unix_time + 60000,
-	}
-
-	if !reflect.DeepEqual(recieved_code, expected_code) {
-		t.Fatalf("Wrong event code info. Expected %v, got %v", expected_code, recieved_code)
-	}
-}
-
-func TestGetEventCodeForbidden(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	id := "testeventid12345"
-	response, err := user_client.New().Get(fmt.Sprintf("/event/code/%s/", id)).ReceiveSuccess(nil)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusForbidden {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestCheckinNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-	CreateProfile()
-	defer ClearProfiles()
-
-	req := event_models.CheckinRequest{
-		Code: "123456",
-	}
-	received_res := event_models.CheckinResult{}
-	response, err := staff_client.New().Post("/event/checkin/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.CheckinResult{
-		NewPoints:   50,
-		TotalPoints: 50,
-		Status:      "Success",
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestCheckinAddToExistingPoints(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-	CreateProfile()
-	defer ClearProfiles()
-
-	client.Database(profile_db_name).Collection("profiles").UpdateOne(
-		context.Background(),
-		bson.M{"id": "theadminprofile"},
-		bson.M{"$set": bson.M{
-			"points": 19,
-		}},
-	)
-
-	req := event_models.CheckinRequest{
-		Code: "123456",
-	}
-	received_res := event_models.CheckinResult{}
-	response, err := staff_client.New().Post("/event/checkin/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.CheckinResult{
-		NewPoints:   50,
-		TotalPoints: 69,
-		Status:      "Success",
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestCheckinInvalidCode(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-	CreateProfile()
-	defer ClearProfiles()
-
-	req := event_models.CheckinRequest{
-		Code: "wrongcode",
-	}
-	received_res := event_models.CheckinResult{}
-	response, err := staff_client.New().Post("/event/checkin/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.CheckinResult{
-		NewPoints:   -1,
-		TotalPoints: -1,
-		Status:      "InvalidCode",
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestCheckinInvalidTime(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-	CreateProfile()
-	defer ClearProfiles()
-
-	req := event_models.CheckinRequest{
-		Code: "abcdef",
-	}
-	received_res := event_models.CheckinResult{}
-	response, err := staff_client.New().Post("/event/checkin/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.CheckinResult{
-		NewPoints:   -1,
-		TotalPoints: -1,
-		Status:      "InvalidTime",
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestCheckinAlreadyCheckedIn(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-	CreateProfile()
-	defer ClearProfiles()
-
-	client.Database(profile_db_name).Collection("profileattendance").UpdateOne(
-		context.Background(),
-		bson.M{"id": "theadminprofile"},
-		bson.M{"$addToSet": bson.M{
-			"events": "testeventid12345",
-		}},
-		options.Update().SetUpsert(true),
-	)
-
-	req := event_models.CheckinRequest{
-		Code: "123456",
-	}
-	received_res := event_models.CheckinResult{}
-	response, err := staff_client.New().Post("/event/checkin/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.CheckinResult{
-		NewPoints:   -1,
-		TotalPoints: -1,
-		Status:      "AlreadyCheckedIn",
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestAddFavoriteEventNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	req := event_models.EventFavoriteModification{
-		EventID: "testeventid12345",
-	}
-	received_res := event_models.EventFavorites{}
-	response, err := user_client.New().Post("/event/favorite/").BodyJSON(req).ReceiveSuccess(&received_res)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_res := event_models.EventFavorites{
-		ID: "localadmin",
-		Events: []string{
-			"testeventid12345",
-		},
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_res, received_res)
-		return
-	}
-
-	res := client.Database(events_db_name).Collection("favorites").FindOne(context.Background(), bson.M{"id": "localadmin"})
-
-	err = res.Decode(&received_res)
-
-	if err != nil {
-		t.Fatalf("Had trouble finding favorites in database: %v", err)
-		return
-	}
-
-	if !reflect.DeepEqual(received_res, expected_res) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", expected_res, received_res)
-	}
-}
-
-func TestAddFavoriteEventNotExist(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	req := event_models.EventFavoriteModification{
-		EventID: "nonexistantid",
-	}
-
-	response, err := user_client.New().Post("/event/favorite/").BodyJSON(req).ReceiveSuccess(nil)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	cursor, err := client.Database(events_db_name).Collection("favorites").Find(context.Background(), bson.M{"id": "localadmin"})
-
-	res := []event_models.EventFavorites{}
-	err = cursor.All(context.TODO(), &res)
-
-	if err != nil {
-		t.Fatalf("Test failed due to unexpected error: %v", err)
-		return
-	}
-
-	expected_res := []event_models.EventFavorites{}
-
-	if !reflect.DeepEqual(res, expected_res) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", expected_res, res)
-	}
-}
-
-func TestGetFavoriteEventsNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	favorite_events := event_models.EventFavorites{
-		ID: "localadmin",
-		Events: []string{
-			"testeventid12345",
-			"testeventid67890",
-		},
-	}
-
-	client.Database(events_db_name).Collection("favorites").InsertOne(context.Background(), favorite_events)
-
-	recieved_favorites := event_models.EventFavorites{}
-	response, err := user_client.New().Get("/event/favorite/").ReceiveSuccess(&recieved_favorites)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	if !reflect.DeepEqual(recieved_favorites, favorite_events) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", favorite_events, recieved_favorites)
-	}
-}
-
-func TestGetFavoriteEventsNone(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	favorite_events := event_models.EventFavorites{
-		ID:     "localadmin",
-		Events: []string{},
-	}
-
-	recieved_favorites := event_models.EventFavorites{}
-	response, err := user_client.New().Get("/event/favorite/").ReceiveSuccess(&recieved_favorites)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	if !reflect.DeepEqual(recieved_favorites, favorite_events) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", favorite_events, recieved_favorites)
-	}
-}
-
-func TestDeleteFavoriteEventsNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	favorite_events := event_models.EventFavorites{
-		ID: "localadmin",
-		Events: []string{
-			"testeventid12345",
-			"testeventid67890",
-		},
-	}
-
-	client.Database(events_db_name).Collection("favorites").InsertOne(context.Background(), favorite_events)
-
-	req := event_models.EventFavoriteModification{
-		EventID: "testeventid67890",
-	}
-	recieved_favorites := event_models.EventFavorites{}
-	response, err := user_client.New().Delete("/event/favorite/").BodyJSON(req).ReceiveSuccess(&recieved_favorites)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_favorites := event_models.EventFavorites{
-		ID: "localadmin",
-		Events: []string{
-			"testeventid12345",
-		},
-	}
-
-	if !reflect.DeepEqual(recieved_favorites, expected_favorites) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_favorites, recieved_favorites)
-	}
-
-	res := client.Database(events_db_name).Collection("favorites").FindOne(context.Background(), bson.M{"id": "localadmin"})
-
-	err = res.Decode(&recieved_favorites)
-
-	if err != nil {
-		t.Fatalf("Had trouble finding favorites in database: %v", err)
-		return
-	}
-
-	if !reflect.DeepEqual(recieved_favorites, expected_favorites) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", expected_favorites, recieved_favorites)
-	}
-}
-
-func TestDeleteFavoriteEventsNone(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	favorite_events := event_models.EventFavorites{
-		ID: "localadmin",
-		Events: []string{
-			"testeventid12345",
-			"testeventid67890",
-		},
-	}
-
-	client.Database(events_db_name).Collection("favorites").InsertOne(context.Background(), favorite_events)
-
-	req := event_models.EventFavoriteModification{
-		EventID: "nonexistantevent",
-	}
-	recieved_favorites := event_models.EventFavorites{}
-	response, err := user_client.New().Delete("/event/favorite/").BodyJSON(req).ReceiveSuccess(&recieved_favorites)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	res := client.Database(events_db_name).Collection("favorites").FindOne(context.Background(), bson.M{"id": "localadmin"})
-
-	err = res.Decode(&recieved_favorites)
-
-	if err != nil {
-		t.Fatalf("Had trouble finding favorites in database: %v", err)
-		return
-	}
-
-	if !reflect.DeepEqual(recieved_favorites, favorite_events) {
-		t.Fatalf("Wrong result received from database. Expected %v, got %v", favorite_events, recieved_favorites)
-	}
-}
-
-func TestGetEventTrackingInfoNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_tracking_users := event_models.EventTracker{}
-	eventid := "testeventid67890"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/event/%s/", eventid)).ReceiveSuccess(&recieved_tracking_users)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_event_tracking := event_models.EventTracker{
-		EventID: eventid,
-		Users: []string{
-			"localadmin",
-		},
-	}
-
-	if !reflect.DeepEqual(recieved_tracking_users, expected_event_tracking) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_event_tracking, recieved_tracking_users)
-	}
-}
-
-func TestGetEventTrackingInfoNone(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_tracking_users := event_models.EventTracker{}
-	eventid := "testeventid12345"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/event/%s/", eventid)).ReceiveSuccess(&recieved_tracking_users)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_event_tracking := event_models.EventTracker{
-		EventID: eventid,
-		Users:   []string{},
-	}
-
-	if !reflect.DeepEqual(recieved_tracking_users, expected_event_tracking) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_event_tracking, recieved_tracking_users)
-	}
-}
-
-func TestGetEventTrackingInfoNonexist(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_tracking_users := event_models.EventTracker{}
-	eventid := "invalideventid"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/event/%s/", eventid)).ReceiveSuccess(&recieved_tracking_users)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
-func TestGetUserTrackingInfoNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_tracked_events := event_models.UserTracker{}
-	userid := "localadmin"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/user/%s/", userid)).ReceiveSuccess(&recieved_tracked_events)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_user_tracking := event_models.UserTracker{
-		UserID: userid,
-		Events: []string{
-			"testeventid67890",
-		},
-	}
-
-	if !reflect.DeepEqual(recieved_tracked_events, expected_user_tracking) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_user_tracking, recieved_tracked_events)
-	}
-}
-
-func TestGetUserTrackingInfoNone(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	recieved_tracked_events := event_models.UserTracker{}
-	userid := "anotheruser"
-	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/user/%s/", userid)).ReceiveSuccess(&recieved_tracked_events)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_user_tracking := event_models.UserTracker{
-		UserID: userid,
-		Events: []string{},
-	}
-
-	if !reflect.DeepEqual(recieved_tracked_events, expected_user_tracking) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_user_tracking, recieved_tracked_events)
-	}
-}
-
-func TestTrackUserNormal(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	checkin_info := checkin_models.UserCheckin{
-		ID:              "localadmin",
-		Override:        false,
-		HasCheckedIn:    true,
-		HasPickedUpSwag: false,
-		RsvpData:        map[string]interface{}{},
-	}
-
-	client.Database(checkin_db_name).Collection("checkins").InsertOne(context.Background(), checkin_info)
-	defer client.Database(checkin_db_name).Collection("checkins").DeleteMany(context.Background(), bson.D{})
-
-	tracking_info := event_models.TrackingInfo{
-		EventID: "testeventid12345",
-		UserID:  "localadmin",
-	}
-
-	res_tracking := event_models.TrackingStatus{}
-	response, err := staff_client.New().Post("/event/track/").BodyJSON(tracking_info).ReceiveSuccess(&res_tracking)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusOK {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-
-	expected_trackers := event_models.TrackingStatus{
-		EventTracker: event_models.EventTracker{
-			EventID: "testeventid12345",
-			Users:   []string{"localadmin"},
-		},
-		UserTracker: event_models.UserTracker{
-			UserID: "localadmin",
-			Events: []string{"testeventid67890", "testeventid12345"},
-		},
-	}
-
-	if !reflect.DeepEqual(res_tracking, expected_trackers) {
-		t.Fatalf("Wrong result received. Expected %v, got %v", expected_trackers, res_tracking)
-	}
-
-	expected_event_tracker := event_models.EventTracker{
-		EventID: "testeventid12345",
-		Users:   []string{"localadmin"},
-	}
-
-	var actual_event_tracker event_models.EventTracker
-
-	res := client.Database(events_db_name).Collection("eventtrackers").FindOne(context.Background(), bson.M{"eventid": "testeventid12345"})
-
-	err = res.Decode(&actual_event_tracker)
-
-	if err != nil {
-		t.Fatalf("Had trouble finding event tracker in database: %v", err)
-		return
-	}
-
-	if !reflect.DeepEqual(actual_event_tracker, expected_event_tracker) {
-		t.Fatalf("Wrong event tracker received from database. Expected %v, got %v", expected_event_tracker, actual_event_tracker)
-	}
-
-	expected_user_tracker := event_models.UserTracker{
-		UserID: "localadmin",
-		Events: []string{"testeventid67890", "testeventid12345"},
-	}
-
-	var actual_user_tracker event_models.UserTracker
-
-	res = client.Database(events_db_name).Collection("usertrackers").FindOne(context.Background(), bson.M{"userid": "localadmin"})
-
-	err = res.Decode(&actual_user_tracker)
-
-	if err != nil {
-		t.Fatalf("Had trouble finding user tracker in database: %v", err)
-		return
-	}
-
-	if !reflect.DeepEqual(actual_user_tracker, expected_user_tracker) {
-		t.Fatalf("Wrong user tracker received from database. Expected %v, got %v", expected_user_tracker, actual_user_tracker)
-	}
-}
-
-func TestTrackUserNotCheckedin(t *testing.T) {
-	CreateEvents()
-	defer ClearEvents()
-
-	tracking_info := event_models.TrackingInfo{
-		EventID: "testeventid12345",
-		UserID:  "localadmin",
-	}
-
-	res_tracking := event_models.TrackingStatus{}
-	response, err := staff_client.New().Post("/event/track/").BodyJSON(tracking_info).ReceiveSuccess(&res_tracking)
-
-	if err != nil {
-		t.Fatal("Unable to make request")
-		return
-	}
-	if response.StatusCode != http.StatusUnprocessableEntity {
-		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
-		return
-	}
-}
-
 func TestStaffActions(t *testing.T) {
+	defer ClearEvents()
 	// 1. Create event
 	event_info := event_models.Event{
 		Name:        "testname",
