+++ tests/e2e/events_test/get_user_tracking_info_test.go
@@ -1,0 +1,66 @@
+package tests
+
+import (
+	"fmt"
+	"net/http"
+	"reflect"
+	"testing"
+
+	"github.com/HackIllinois/api/services/event/models"
+)
+
+func TestGetUserTrackingInfoNormal(t *testing.T) {
+	CreateEvents()
+	defer ClearEvents()
+
+	recieved_tracked_events := models.UserTracker{}
+	userid := "localadmin"
+	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/user/%s/", userid)).ReceiveSuccess(&recieved_tracked_events)
+
+	if err != nil {
+		t.Fatal("Unable to make request")
+		return
+	}
+	if response.StatusCode != http.StatusOK {
+		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
+		return
+	}
+
+	expected_user_tracking := models.UserTracker{
+		UserID: userid,
+		Events: []string{
+			"testeventid67890",
+		},
+	}
+
+	if !reflect.DeepEqual(recieved_tracked_events, expected_user_tracking) {
+		t.Fatalf("Wrong result received. Expected %v, got %v", expected_user_tracking, recieved_tracked_events)
+	}
+}
+
+func TestGetUserTrackingInfoNone(t *testing.T) {
+	CreateEvents()
+	defer ClearEvents()
+
+	recieved_tracked_events := models.UserTracker{}
+	userid := "anotheruser"
+	response, err := staff_client.New().Get(fmt.Sprintf("/event/track/user/%s/", userid)).ReceiveSuccess(&recieved_tracked_events)
+
+	if err != nil {
+		t.Fatal("Unable to make request")
+		return
+	}
+	if response.StatusCode != http.StatusOK {
+		t.Fatalf("Request returned HTTP error %d", response.StatusCode)
+		return
+	}
+
+	expected_user_tracking := models.UserTracker{
+		UserID: userid,
+		Events: []string{},
+	}
+
+	if !reflect.DeepEqual(recieved_tracked_events, expected_user_tracking) {
+		t.Fatalf("Wrong result received. Expected %v, got %v", expected_user_tracking, recieved_tracked_events)
+	}
+}
